MINOR: move RecordReader from org.apache.kafka.tools to org.apache.common
