+++ core/src/main/scala/kafka/tools/ConsoleProducer.scala
@@ -27,11 +27,10 @@ import kafka.utils.Implicits._
 import kafka.utils.{Exit, Logging, ToolsUtils}
 import org.apache.kafka.clients.producer.internals.ErrorLoggingCallback
 import org.apache.kafka.clients.producer.{KafkaProducer, Producer, ProducerConfig, ProducerRecord}
-import org.apache.kafka.common.KafkaException
+import org.apache.kafka.common.{KafkaException, RecordReader}
 import org.apache.kafka.common.record.CompressionType
 import org.apache.kafka.common.utils.Utils
 import org.apache.kafka.server.util.{CommandDefaultOptions, CommandLineUtils}
-import org.apache.kafka.tools.RecordReader
 
 import scala.annotation.nowarn
 
@@ -45,7 +44,7 @@ object ConsoleProducer extends Logging {
         r.configure(prop.asInstanceOf[java.util.Map[String, _]])
         r
       case r: MessageReader =>
-        logger.warn("MessageReader is deprecated. Please use org.apache.kafka.tools.RecordReader instead")
+        logger.warn("MessageReader is deprecated. Please use org.apache.kafka.common.RecordReader instead")
         new RecordReader {
           private[this] var initialized = false
